Complete FastAPI masterclass plus Instagram clone project PC

+++ Udemy - Complete FastAPI masterclass plus Instagram clone project/db/db_user.py
@@ -1,3 +1,4 @@
+from fastapi import HTTPException, status
 from db.hash import Hash
 from sqlalchemy.orm.session import Session
 from schemas import UserBase
@@ -5,34 +6,46 @@ from db.models import DbUser
 
 
 def create_user(db: Session, request: UserBase):
-  new_user = DbUser(
-    username = request.username,
-    email = request.email,
-    password = Hash.bcrypt(request.password)
-  )
-  db.add(new_user)
-  db.commit()
-  db.refresh(new_user)
-  return new_user
+    new_user = DbUser(
+        username=request.username,
+        email=request.email,
+        password=Hash.bcrypt(request.password)
+    )
+    db.add(new_user)
+    db.commit()
+    db.refresh(new_user)
+    return new_user
+
 
 def get_all_users(db: Session):
-  return db.query(DbUser).all()
+    return db.query(DbUser).all()
+
 
 def get_user(db: Session, id: int):
-  return db.query(DbUser).filter(DbUser.id == id).first()
+    user = db.query(DbUser).filter(DbUser.id == id).first()
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f'User with id {id} not found')
+    return db.query(DbUser).filter(DbUser.id == id).first()
+
 
 def update_user(db: Session, id: int, request: UserBase):
-  user = db.query(DbUser).filter(DbUser.id == id)
-  user.update({
-    DbUser.username: request.username,
-    DbUser.email: request.email,
-    DbUser.password: Hash.bcrypt(request.password)
-  })
-  db.commit()
-  return 'ok'
+    user = db.query(DbUser).filter(DbUser.id == id)
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f'User with id {id} not found')
+    user.update({
+        DbUser.username: request.username,
+        DbUser.email: request.email,
+        DbUser.password: Hash.bcrypt(request.password)
+    })
+    db.commit()
+    return 'ok'
+
 
 def delete_user(db: Session, id: int):
-  user = db.query(DbUser).filter(DbUser.id == id).first()
-  db.delete(user)
-  db.commit()
-  return 'ok'
+    user = db.query(DbUser).filter(DbUser.id == id).first()
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f'User with id {id} not found')
+    db.delete(user)
+    db.commit()
+    return 'ok'
+# raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f'Article with id {id} not found')
